# emergency-rescue-planner/service/go/bff/docs/swagger.yaml
basePath: /
definitions:
...
  route_engine.GeoJSONGeometry:
    properties:
      coordinates:
        items:
          items:
            format: float64
            type: number
          type: array
        type: array
      type:
        type: string
    type: object
  route_engine.LatLng:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  route_engine.RouteData:
    properties:
      bbox:
        items:
          type: number
        type: array
      distance_metres:
        type: number
      duration_seconds:
        type: number
      geometry:
        $ref: '#/definitions/route_engine.GeoJSONGeometry'
      steps:
        items:
          $ref: '#/definitions/route_engine.RouteStep'
        type: array
    type: object
  route_engine.RouteRequest:
    properties:
      avoid_zone_ids:
        items:
          type: string
        type: array
      destination:
        $ref: '#/definitions/route_engine.LatLng'
      origin:
        $ref: '#/definitions/route_engine.LatLng'
      transport:
        type: string
      waypoints:
        items:
          $ref: '#/definitions/route_engine.LatLng'
        type: array
    type: object
  route_engine.RouteResponse:
    properties:
      data:
        $ref: '#/definitions/route_engine.RouteResponseData'
    type: object
  route_engine.RouteResponseData:
    properties:
      affected_zone_ids:
        items:
          type: string
        type: array
      route:
        $ref: '#/definitions/route_engine.RouteData'
    type: object
  route_engine.RouteStep:
    properties:
      distance_m:
        type: number
      duration_sec:
        type: number
      instruction:
        type: string
      maneuver_type:
        type: string
      way_point_index:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the emergency rescue planner
  title: Emergency Rescue Planner API
  version: "1.0"
paths:
...
  /api/route/calculate:
    post:
      consumes:
      - application/json
      description: Returns a road-network route with turn-by-turn steps. Requires
        ORS_API_KEY.
      parameters:
      - description: Route Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/route_engine.RouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/route_engine.RouteResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "503":
          description: routing service unavailable
          schema:
            type: string
      summary: Calculate a road-network route
      tags:
      - routing
...
swagger: "2.0"